Calculator functions fix

+++ Calculator/Addition.py
@@ -1,2 +1,2 @@
 def addition(a, b):
-    return int(b) + int(a)
+    return float(b) + float(a)
