+++ Calculator/Square.py
@@ -1,2 +1,2 @@
 def square(a):
-    return int(a) * int(a)
+    return float(a) * float(a)
